+++ test/commands/changes.js
@@ -17,7 +17,7 @@ const commands = [
 ];
 
 for (const command of commands) {
-    test(`test "${command}" command. Precondition: CHANGELOG.md exists. Does not contain any new change. / Postcondition: A new change is set as ${command} to the CHANGELOG.md`, (t) => {
+    test(`test "${command}" command => CHANGELOG.md exists. Does not contain any new change. => A new change is set as ${command} to the CHANGELOG.md`, (t) => {
         return Promise.all([cli(tmp, { name: command, args: { msg: 'super *cool feature*' } }, 'changelog_exists'), readChangelog(`expected/${command}/changelog_exists`)])
             .then((values) => {
                 let [result, expected] = values;
@@ -25,28 +25,17 @@ for (const command of commands) {
             });
     });
 
-    test(`test "${command}" command. Precondition: CHANGELOG.md already exists and contains changes. / Postcondition: A new change is set as ${command} to the CHANGELOG.md, previous changes are maintaned ok.`, (t) => {
+    test(`test "${command}" command => CHANGELOG.md already exists and contains changes. => A new change is set as ${command} to the CHANGELOG.md, previous changes are maintaned ok.`, (t) => {
         return Promise.all([cli(tmp, { name: command, args: { msg: 'super *cool feature*' } }, 'changelog_with_items'), readChangelog(`expected/${command}/changelog_with_items`)])
             .then((values) => {
                 let [result, expected] = values;
                 t.deepEqual(result, expected, `chan injected a new change labeled as ${command} to an already populated CHANGELOG.md.`);
             });
     });
-
-    //
-    // we have to change this test, because now if there is no user input Chan opens the editor by default.
-    //
-    // test(`test "${command}" command. Precondition: CHANGELOG.md already exists but there is no user input. / Postcondition: CHANGELOG.md remains the same.`, (t) => {
-    //     return Promise.all([cli(tmp, { name: command }, 'changelog_exists'), readChangelog(`fixtures/${command}/changelog_exists`)])
-    //         .then((values) => {
-    //             let [result, expected] = values;
-    //             t.deepEqual(result, expected, 'chan does not modify CHANGELOG.md.');
-    //         });
-    // });
 }
 
 
-test('test 2 commands in a row. Precondition: CHANGELOG.md already exists and contains changes. / Postcondition: CHANGELOG.md is fullfiled according the previous commands.', (t) => {
+test('test 2 commands in a row => CHANGELOG.md already exists and contains changes. => CHANGELOG.md is fullfiled according the previous commands.', (t) => {
     const syncActions = cli(tmp, [
         { name: 'added', args: { msg: 'added 1' } },
         { name: 'added', args: { msg: 'added 2' } },