+++ test/commands/expected/release/unreleased_yanked/CHANGELOG.md
@@ -5,3 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [1.0.1] - <currentDate> [YANKED]
+
+## 1.0.0 - 2016-06-18
+### Added
+- test change 1
+
+[unreleased]: https://github.com/geut/chan/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/geut/chan/compare/v1.0.0...v1.0.1