+++ test/commands/fixtures/release/unreleased_yanked/CHANGELOG.md
@@ -5,3 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## 1.0.0 - 2016-06-18
+### Added
+- test change 1
+
+[unreleased]: https://github.com/geut/chan/compare/v1.0.0...HEAD